refactored dashes formatting provider and dispatcher
